백준 9782: Median
The median is the middle value of an ordered group of data. That is, the amount of data is less and more than the median. The base will have the same number.

Let d1, d2, d3, ..., dn be all data.

If n is odd, the median is the position data at (n+1)/2, that is, Median = d(n+1)/2.

But if n is even, the median is the mean of (n/2) and (n/2)+1 position data, that is, Median = (d(n/2) + d(n/2)+1)/2.