백준 27257: Любитель нулей(Zero Lover)
Sasha loves zeros very much. But the zeros at the end of the number do not seem interesting to him. Of course, Sasha is not interested in leading zeros either.

Sasha considers the beauty of a number to be equal to the number of zeros in its decimal notation, excluding zeros at the end of the number. Of course, there should be no leading zeros in the entry. For example, the beauty of the number 100500 is 2.

Given a number k, find out what his beauty is according to Sasha.