백준 27590: Sun and Moon
You recently missed an eclipse and are waiting for the next one! To see any eclipse from you home, the sun and the moon must be in alignment at specific positions. You know how many years ago the sun was in the right position, and how many years it takes for it to get back to that position. You know the same for the moon. When will you see the next eclipse?