백준 5666: Hot Dogs
In 2012 a new world record was set in the famous Nathan’s Hot Dog Eating Competition: the champion, Joey Chestnut, ate 68 hot dogs in ten minutes, an amazing increase from the 62 hot dogs eaten by the same Chestnut in 2011.

Nathan’s Famous Corporation, in Brooklyn, NY, is responsible for the contest. They make delicious hot dogs, famous worldwide, but when it comes to math they are not so good. They want to apply for being listed in the Guinness World of Records, but for that they need to fill up a form describing basic facts about the competition. In particular, they are required to inform the average number of hot dogs eaten by participants during the contest.

Can you help them? They promised to pay you with one of their tasty hot dogs. Given the total number of hot dogs consumed and the total number of participants in the contest, you must write a program to determine the average number of hot dogs eaten by participants.