백준 5235: Even Sum More Than Odd Sum
When new programs arrive in the grid world, they start by playing the simplest of games in the Disc Arena against other novice programs. One of those games is played in front of a large board as follows: a sequence of numbers appears on the board, and the players have to decide whether the sum of even numbers in the sequence is larger than the sum of odd numbers in the sequence, or if they are equal. The first player to provide the correct answer wins the game, and the loses are deleted.

For example, given the sequence:

5 5 2 1 10 13 6 12

the sum of even numbers is 2 + 10 + 6 + 12 = 30, whereas the sum of odd numbers is 5 + 7 + 1 + 13 = 26. Hence the correct answer here is "EVEN".

You are to help Sam (our protagonist) win this game so he can survive, and look for his father.