백준 28249: Chili Peppers
Ron is cooking chili using an assortment of peppers.

The spiciness of a pepper is measured in Scoville Heat Units (SHU). Ron's chili is currently not spicy at all, but each time Ron adds a pepper, the total spiciness of the chili increases by the SHU value of that pepper.

The SHU values of the peppers available to Ron are shown in the following table:

Pepper Name	Scoville Heat Units
Poblano		1500
Mirasol		6000
Serrano		15500
Cayenne		40000
Thai		75000
Habanero	125000

Your job is to determine the total spiciness of Ron's chili after he has fiished adding peppers.