백준 2387: Howl
On a full moon, one might be able to hear a wolf howl in the distance. You manage to come across a wolf howling at the moon, and you want to capture the beautiful image. Unfortunately, you left your camera at home, so you decide to just create the picture on your computer! Create the image of a wolf howling as shown below.