백준 28490: Area
Stuart has n rectangular frames, which are numbered from 1 to n. Frame i is a rectangle with height h[i] and width w[i].

The size of a frame is the area that it covers. Stuart wants you to help him find the area covered by the largest size frame that he has.