백준 29340: Отряд(Squad)
The Rebels attack the Empire again on the planet Hoth, and Darth Vader decides to send his best clones there. But here's the problem: there is only one starship, and there are two detachments to control it. Then he decided to combine these squads so that only the best clones remained in the resulting squad.

Both squads consist of the same number of clones. Each clone in a squad has its own specialty, and in the same squad there are no two clones with the same specialties. It is clear that the set of specialties of clones in the first squad coincides with the set of specialties in the second.

To select the best clones, a special card was made for each squad with all the clone scores they once received at the academy. It is known that the cards were made as follows:

- another clone was called from the first squad
- a clone was called from the second detachment, owning the same specialty as the clone from the first detachment
- the score of the clone from the first squad was added to the right on the card of the first squad, and the score of the clone from the second squad was added to the card of the second squad
- the described actions were performed as long as there were unaccounted for clones in the units
- clone score --- number from 0 to 9

After that, for each specialty, the clone whose score was no less than that of a clone with a similar specialty was selected. At the end, a simial card was entered for the resulting detachment. It is known that when compiling it, the order of specialties was the same as when compiling cards for the two original detachments. However, this card was lost, and now you need help to restore it.