백준 28808: Таблица результатов(Results Table)
Ralph and Vanilope, while traveling through the games, found themselves on the computer of a participant of a programming Olympiad. Moreover, they found out that the participant was offered n tasks and each of them can be sent a solution no more than m times.

The heroes also found a table with the results of testing the solutions. The table consists of n rows and m columns. In the cell, which is in the i-th row and j-th column is written the symbol:

- <<+>>, if the j-th sent solution for the i-th problem is correct.
- <<->>, if the j-th sent solution for the i-th task is incorrect.
- <<.>>, if there were no j submitted solutions for the i-th problem.

Vanilope explained to Ralph that the first correct solution for each problem is counted and all other solutions sent for that problem do not affect the participant's result.

Ralph wondered how many of the problems the contestant was able to solve. But since he is too shy to ask Vanilope for help, help him!