백준 29546: Files
Sergey is fond of digital photography. Recently he decided to print out some of his photos and hang them on the wall in the hallway. To do this, he wrote out the file names of all the photos and numbered them in the order in which they appear on the list.

After that, Sergey spent several hours selecting the best photos. Usually Sergey liked several photos in a row at once, so he wrote out the number of the first and the last of them. As a result, he had a list of intervals of the numbers of the photos he liked.

Now Sergey wants to get a list of file names of all the photos he likes. This list should first list the file names of photos from the first interval in ascending order, then from the second interval, and so on.

This list can be very large, so Sergey asks you to help him.