백준 29823: Pakirobot Manhattanis
A self-driving parcel robot is operating in the Manhattan borough of New York. This robot is quite large and can deliver several parcels to different destinations at once. After delivering all the parcels, the robot has to go back to the warehouse to pick up new ones.

The streets of Manhattan from a grid: all the streets are exactly north-south or east-west. In addition, all adjacent parallel streets have equal distances.

Let's call a robot's step from one intersection to the next intersection along any street.

Find the number of steps it takes the robot to get back to the warehouse.