백준 29986: Altura Mínima(Minimum Height)
Carlitos is an adventure enthusiast with an insatiable love of amusement parks. Despite his vibrant passion, Carlitos faces a unique challenge: his limited height. As he anxiously plans his weekend adventure, he realizes that his vertical limitations may get in the way of his amusement park experience. It's not just about choosing a park; it's about finding one where he can enjoy the thrill of the rides.

Imagine the kaleidoscope of colors, the jubilant laughter and the adrenaline rush of the rides. Carlitos has always been attracted to the energy of amusement parks. As the weekend approaches, he pores over the park brochures, studying the height requirements of each ride. His goal is to maximize your enjoyment, and that's where you come in.

Your task is to help Carlitos determine the number of rides he can enjoy in a specific park. Considering his height and the minimum height requirements of each ride, guide him to make the most of his amusement park adventure.