백준 28938: Конвейер(Conveyor)
Today Dipper discovered a mysterious conveyor belt in the woods. Of course, he found a page in his diary that contains some information about this object. For example, he noticed a sequence of numbers, each of which is equal to -1, 0 or 1.

Next to the conveyor belt, he found a control panel with three buttons with the same numbers written on them. He had already checked what happens when you press them. It turned out that when he pressed <<-1>>, the conveyor moved one inch to the left, when he pressed <<1>>, the conveyor moved one inch to the right, and when he pressed <<0>>, there was no visible change (although, who knows, maybe a portal to a parallel world was opened somewhere).

Dipper has already set himself to press the sequence of numbers from the diary, but before that he decided to be cautious and find out at least which way the conveyor will eventually move from its current position.

Help him out.