백준 30067: Paslėpta suma (Hidden Amount)
Simas and George play this game. Simas comes up with nine integers, positive and nebative, for example: 5, 4, -5, -3, 9, 1, 2, 5, -11. He then counts up the sum of 5 + 4 + (-5) + (-3) + 9 + 1 + 2 + 5 + (-11) = 7, and hides it somewhere in between these numbers. This gives a total of ten numbers:

5, 4, -5, -3, 9, 1, 7, 2, 5, -11

Simas then shows this sequence to George. George's goal is to find which of the ten numbers is the hidden sum of the remaining numbers as quickly as possible.

Write a program to help George find the hiddem sum.