백준 28960: Плащ левитации (Cloak of Levitation)
While fighting another demon, Doctor Strange heavily soiled his irreplaceable cloak of levitation. But since using magic without any special need is bad manners, Stephen will do laundry in the machine like an ordinary mortal.

However, it turned out that washing is far from the most difficult dilemma on the agenda. Strange wants to hang his coat to dry, but all he has in this dimension is a clothesline stretched at a height of h centimeters and a length of l centimeters. The cloak is a rectangle with sides a and b centimeters.

Stephen wants to throw the cloak over the rope, but it is very important that the hanging cloak does not touch the floor. Stephen was lucky that the rope was inextensible and would maintain its height even with a cloak hanging on it. Also, if the cloak hangs exactly to the floor, the touch does not count, since it is very thin and the area of contact in this case is negligible.

The doctor's hands are busy with the cloak, and his mind is full of great thoughts, so he asks you to tell him if he can hang the cloak on the rope without it touching the floor.