백준 28940: Дневник Гравити Фолз(Gravity Falls Diary)
One day, walking through the forest, Dipper found a new diary, and he was overjoyed. But his happiness was short-lived: when he opened the cover, he saw only scraps of parchment with letters in it---someone probably didn't want the diary to be read.

Dipper knows that a diary in Gravity Falls is usually a lot of pages filled out in a special way. All the letters in the diary are the same size, that is, the same width and height. The horizontal and vertical spacing between these letters is so small that it can be considered zero. In addition, on each page there is a solid text. That is, the text is organized as follows: if a letter can be written in a line (i.e. it fits in width), it is written in the current line. If the letter does not fit on the current line, it is written on the next line (which is lower). If there is no next line, for example, because the letter does not fit on the current page, then the letter is moved to the next page in the upper left corner and the whole thing is repeated again.

Dipper now knows the total number of letters, as well as the size of the diary pages and the size of the letters. Knowing this, Dipper can easily count how many pages there were in the diary, assuming there are no blank pages without text. Can you?