백준 29683: Рождественская лотерея(Christmas Raffle)
Christmas is just around the corner, but the stores of the AKEI retail chain are still empty. To attract buyers, sales manager Anton came up with the following promotion. For every A ruble that the buyer spends, he receives one lottery ticket, or, as the advertisement says, a chance to win a car.

However, not all cashiers were aware of this promotion, and some buyers were left without the tickets they were entitled to. Now Anton receives a pile of checks every day, according to which he must be given lottery tickets. He hasn't slept for three nights and is very tired.

Help him figure out how many tickets need to be issued.