백준 30664: Loteria Falha(Failed Lottery)
Many people dream of making easy money. Some people try to do this through the lottery. They buy games such as "scratch cards" and lottery tickets in anticipation of multimillion-dollar draws.

Gilmar, a very naughty boy, decided to use his mathematical knowledge to try to increase his chances in these games of chance. He bought several tickets of the same type of scratch card and analyzed the cards until he noticed a very curious property: each scratch card had an identifying number printed on it which gave him an idea of the chances of winning.

Out of every 10 cards he bought at the stall, approximately 1 or 2 came with some kind of prize: at least another free scratch card, or a symbolic cash prize. When he applied his method to choosing which cards to buy, he realized that out of every 10 cards, on average 8 contained a prize!

Since the task is too tiring to do manually, he thought that you, his long-time friend, could help him with a program that, given the card's identifying number, tells him if it is part of the set of cards with the highest chance of winning. The trick is to look for cards whose identifying number is a multiple of 42. Are you able to help your colleague in this endeavor?