백준 32326: Conveyor Belt Sushi
There is a new conveyor belt sushi restaurant in town. Plates of sushi travel around the restaurant on a raised conveyor belt and customers choose what to eat by removing plates.

Each red plate of sushi costs $3, each green plate of sushi costs $4, and each blue plate of sushi costs $5.

Your job is to determine the cost of a meal, given the number of plates of each colour chosen by a customer.